requirements.txt
PyQt6
Pillow
adif-io